% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Creates a grid that can be filled in using the function grid_search}
\usage{
create_grid(x, y, method)
}
\arguments{
\item{x}{Matrix or data frame containing the dependent variables..}

\item{y}{Vector of responses. Can either be a factor or a numeric vector.}

\item{method}{Model to be fit. Choices are "ada" for adaboost, "en" for
elastic net, "gbm" for gradient boosting machines, and "svm" for support
vector machines.}
}
\value{
Returns a dataset that forms the grid for the search. It contains
the data frame has a column for each of the tuning parameters and a
column for the Accuracy/MSE, AUC/MAE, and one for the time in seconds.
}
\description{
create_grid creates a grid that can be filled in using the function
grid_search. It is kept separate from the grid_search functions so that
the user can create a different grid if they choose.
}
\seealso{
\code{\link{grid_search}}
}
