% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_benchmark.R
\name{en_benchmark}
\alias{en_benchmark}
\title{Tunes a model via elastic net with \code{glmnet} and returns}
\usage{
en_benchmark(x, y, name = "Data", loss = "default")
}
\arguments{
\item{x}{Matrix or data frame containing the dependent variables.}

\item{y}{Vector of responses. Can either be a factor or a numeric vector.}

\item{name}{Name of the dataset. Used to name the output file and
as an identifier within the output dataset.}

\item{loss}{Value for \code{loss} argument in \code{glmnet}. Options are
"default", "class", "auc", "mse", or "mae". If "default" is selected,}
}
\value{
Saves a matrix to the indicated path that contains the results
for each of the runs. The final file contains the following variables:
\item{data}{Name of the dataset.}
\item{package}{This will always be "glmnet".}
\item{method}{Type of model that was fit. This will always be "en"
for elastic net.}
\item{optimizer}{Optimization method. It will be "1se" or "min" to
reflect the criteria used for lambda selection by glmnet.}
\item{assess}{The method used to assess the model during tuning. It will
always be cross-validation.}
\item{tuned_on}{Metric that was used as the loss function.}
\item{acc_rmse}{The best accuracy or RMSE obtained from the model as
determined by a validation dataset.}
\item{auc_mae}{The best AUC or MAE obtained from the model as
determined by a validation dataset.}
\item{time}{Time to complete the calculations in seconds.}
The models are verified using rsample to split the data into  training
and testing datasets.
}
\description{
\code{en_benchmark} tunes a model with \code{glmnet} and returns a one row
\code{data.frame} with the results. \pkg{rsample} is used to split
the dataset into training and test datasets. The results returned
}
\seealso{
\code{\link{tm_benchmark}}, \code{\link{EZtune::eztune}}
}
