% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztune_table.R
\name{eztune_table}
\alias{eztune_table}
\title{Makes a table of performance results}
\usage{
eztune_table(dat, value = "loss_mse_acc_10", best_grid = NULL)
}
\arguments{
\item{dat}{Dataset created from load_opt_data or with several
datasets from load_opt_data combined with rbind. Information
from this dataset will be used to populate the majority of
the table.}

\item{value}{the value from dat that should be used as the
error metric. This will typically be either loss_mse_acc_10 or
loss_mae_auc_10}

\item{best_grid}{A dataset that contains the best error rate
obtained from the grid search. The first column must contain
the names of the datasets used to populate the tables. The
names must be identical to those in the dat object. The second
column is numeric and contains the best error measure obtained
by the grid search for that dataset. If it is omitted the table will
be made without the grid information. The function get_best_grid
can be used to create}
}
\value{
Returns a data.frame with the performance metrics from the
EZtune tests. Each column is a character vector. This data.frame
can be passed to the function eztune_table_latex to create a latex
version of the table.
}
\description{
eztune_table creates a table that contains the average error rates
and run times for a series of eztune runs and compares them to
the best results obtained from a grid search.
}
\seealso{
\code{\link{load_opt_data}}, \code{\link{get_best_grid}},
\code{\link{get_grid_data}}, \code{\link{eztune_table_latex}}
}
