% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztune_table_latex.R
\name{eztune_table_latex}
\alias{eztune_table_latex}
\title{Makes a table that can be pasted directly into latex}
\usage{
eztune_table_latex(tab, vert_lines = TRUE, sideways = TRUE)
}
\arguments{
\item{tab}{A table created with eztune_table.}

\item{vert_lines}{If TRUE, vertical lines will be placed between the
columns. If FALSE, no vertical lines will appear in the table.}
}
\value{
Returns a character vector that contains the header and the
lines for a latex table version of an eztune_table. Commands to
end the table and horizontal line commands need to be added later.
}
\description{
eztune_table_latex creates a table created with eztune_table
that can be pasted directly into latex. Only the horizontal
lines need to be added.
}
\seealso{
\code{\link{eztune_table}}
}
