% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_grid.R
\name{get_best_grid}
\alias{get_best_grid}
\title{Makes a data.frame that contains the best error rates from a
grid search}
\usage{
get_best_grid(grid_dat)
}
\arguments{
\item{grid_dat}{data.frame obtained from get_grid_data or with
several datasets from get_grid_data combined with rbind.}
}
\value{
Returns a data.frame with the names of the datasets
in the first column and the best error rate in the second
column. The first column is named "Data" and the second column
is named "Best"
}
\description{
get_best_grid creates a data.frame that has the datasets in
the first column and the best error rate obtained in the grid
search in the second column.
}
\seealso{
\code{\link{get_grid_data}}, \code{\link{eztune_table}},
\code{\link{grid_search}},
}
