% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_plot.R
\name{opt_plot}
\alias{opt_plot}
\title{Makes a line plot of the results from the optimization tests}
\usage{
opt_plot(dat, value, facet = "optimizer", cols = NULL, title = NULL)
}
\arguments{
\item{dat}{Dataset created from load_opt_data or with several
datasets from load_opt_data combined with rbind.}

\item{value}{Specifies which error metric should be plotted.
This should be the name of a variable in the dataset, dat.}

\item{facet}{Specifies a faceting variable. The typical faceting
variable is optimizer, but other options include data or method.}

\item{cols}{A color palette to use with the data. If it is not
specified, the Dark2 palette from color brewer will be used.}

\item{title}{Optional character string to use a title for the plot.}
}
\value{
Returns a parallel coordinate plot created by ggplot.
The x-axis shows the different optimization methods used, the
y-axis shows the standardized error rate for each dataset.
Within each dataset, the optimization method with the largest
loss is assigned a value of 1 and the smallest is assigned a
value of 0. If a faceting variable is selected, standardization
will be done regardless of the faceted variable. Standardizing is
done within each dataset so each dataset should have a value of 1
and a value of 0 on the graph. Datasets that are all NA for an
optimization method are denoted with a star and a value
of 1.
}
\description{
opt_plot creates a line plot with the results of the optimization
tests.
}
\seealso{
\code{\link{load_opt_data}}, \code{\link{average_metric}}
}
