% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_benchmark.R
\name{tm_benchmark}
\alias{tm_benchmark}
\title{Tunes a model via tidymodels}
\usage{
tm_benchmark(
  x,
  y,
  name = "Data",
  method = "svm",
  test = "grid",
  grid_size = 5,
  bayes_iter = 10
)
}
\arguments{
\item{x}{Matrix or data frame containing the dependent variables.}

\item{y}{Vector of responses. Can either be a factor or a numeric vector.}

\item{name}{Name of the dataset. Used to name the output file and
as an identifier within the output dataset.}

\item{method}{Model to be fit. Choices are "ada" for adaboost, "en" for
elastic net, "gbm" for gradient boosting machines, and "svm" for support
vector machines.}

\item{test}{Type of tuning to be done. Choices are "bayes" for iterative
Bayesian optimization and "grid" for grid search.}

\item{grid_size}{Number of values for each hyperparameter for tidymodels
grid. This is ignored if iterative Bayesian optimization is selected.}

\item{bayes_iter}{Number of iterations for the Bayesian optimization.
This is ignored if test = "grid"}
}
\value{

}
\description{
tm_benchmark tunes a model with tidymodels and returns the accuracy
results and computation time in seconds.
}
\seealso{
\code{\link{ez_benchmark}}, \code{\link{EZtune::eztune}}
}
