% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztune_results.R
\name{eztune_results}
\alias{eztune_results}
\title{Performs a specified number of eztune results}
\usage{
eztune_results(
  x,
  y,
  data_name,
  method = NULL,
  optimizer = NULL,
  fast = NULL,
  cross = NULL,
  loss = NULL,
  iterations = 10,
  path = "."
)
}
\arguments{
\item{x}{Matrix or data frame containing the dependent variables.}

\item{y}{Vector of responses. Can either be a factor or a numeric vector.}

\item{data_name}{Name of the dataset. Used to name the output file and
as an identifier within the output dataset.}

\item{method}{Model to be fit. Choices are "ada" for adaboost, "en" for
elastic net, "gbm" for gradient boosting machines, and "svm" for support
vector machines.}

\item{optimizer}{Optimization method. Options are "ga" for a genetic
algorithm and "hjn" for a Hooke-Jeeves optimizer.}

\item{fast}{Indicates if the function should use a subset of the
observations when optimizing to speed up calculation time. A value of
TRUE will use the smaller of 50% of the data or 200 observations for
model fitting, a number between 0 and 1 specifies the proportion of
data to be used to fit the model, and a postive integer specifies the
number of observations to be used to fit the model. A model is computed
using a random selection of data and the remaining data are used to
validate model performance. The validation error measure is used as
the optimization criterion.}

\item{cross}{If an integer k > 1 is specified, k-fold cross-validation
is used to fit the model. This method is very slow for large datasets.
This parameter is ignored unless fast = FALSE.}

\item{loss}{The type of loss function used for optimization. Options for
models with a binary response are "class" for classification error and
"auc" for area under the curve. Options for models with a continuous
response are "mse" for mean squared error and "mae" for mean absolute
error. If the option "default" is selected, or no loss is specified,
the classification accuracy will be used for a binary response model
and the MSE will be use for models with a continuous model.}

\item{iterations}{Number of times to run the model.}

\item{path}{Where the file should be saved.}
}
\value{
Returns a parallel coordinate plot created by ggplot.
The x-axis shows the different optimization methods used, the
y-axis shows the standardized error rate for each dataset.
Within each dataset, the optimization method with the largest
loss is assigned a value of 1 and the smallest is assigned a
value of 0. If a faceting variable is selected, standardization
will be done regardless of the faceted variable. Standardizing is
done within each dataset so each dataset should have a value of 1
and a value of 0 on the graph. Datasets that are all NA for an
optimization method are denoted with a hollow circle and a value
of 1.
}
\description{
eztune_results creates a line plot with the results of the optimization
tests.
}
\seealso{
\code{\link{load_opt_data}}, \code{\link{average_metric}}
}
